# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/verifications": {
      "post": {
        "operationId": "VerificationsController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVerificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create and enqueue a verification",
        "tags": [
          "Verifications"
        ]
      },
      "get": {
        "operationId": "VerificationsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "QUEUED",
                "PROCESSING",
                "PARTIALLY_COMPLETED",
                "COMPLETED",
                "FAILED",
                "CANCEL_REQUESTED",
                "CANCELLED",
                "DELETED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Verifications"
        ]
      }
    },
...
    "/api/v1/verifications/{id}": {
      "get": {
        "operationId": "VerificationsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Verifications"
        ]
      },
      "delete": {
        "operationId": "VerificationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Verifications"
        ]
      }
    },
...
    "/api/v1/verifications/{id}/retry": {
      "post": {
        "operationId": "VerificationsController_retry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Verifications"
        ]
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search users without exposing credentials",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get safe administrative user details",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}/status": {
      "patch": {
        "operationId": "AdminController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change account status and revoke sessions when required",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}/role": {
      "patch": {
        "operationId": "AdminController_changeRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change a user role as a super administrator",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/verifications": {
      "get": {
        "operationId": "AdminController_listVerifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inspect verification lifecycle metadata",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/verifications/{id}": {
      "get": {
        "operationId": "AdminController_getVerification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inspect safe verification lifecycle metadata without input",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/verifications/{id}/retry": {
      "post": {
        "operationId": "AdminController_retryVerification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Idempotently enqueue a failed verification retry",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "operationId": "AdminController_listAuditLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List append-only administrative audit records",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Verith API",
    "description": "Explainable misinformation verification and media literacy API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "verith_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "verith_refresh"
      }
    },
    "schemas": {
...
      "CreateVerificationDto": {
        "type": "object",
        "properties": {
          "sourceType": {
            "type": "string",
            "enum": [
              "TEXT",
              "URL",
              "IMAGE",
              "SCREENSHOT",
              "AUDIO",
              "WHATSAPP_TEXT",
              "WHATSAPP_URL",
              "WHATSAPP_IMAGE",
              "WHATSAPP_AUDIO"
            ]
          },
          "text": {
            "type": "string",
            "maxLength": 50000
          },
          "url": {
            "type": "string"
          },
          "mediaAssetId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "question": {
            "type": "string",
            "maxLength": 1000
          },
          "requestedLanguage": {
            "type": "string",
            "default": "en"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "UNLISTED",
              "PUBLIC"
            ],
            "default": "PRIVATE"
          }
        },
        "required": [
          "sourceType"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "AdminUserStatusDto": {
        "type": "object",
        "properties": {}
      },
      "AdminUserRoleDto": {
        "type": "object",
        "properties": {}
      },
      "AdminReasonDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}